+++ p2/headers/aux.hpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 void rellenarVector(std::vector<int> &v);
-void printVector(std::vector<int> &v);
+void printVector(std::vector<double> &v);
 bool estaOrdenado(const std::vector <int> &v);
 int getMedian(int n, const std::vector <int> &v);
 void printHelp();
